// main/src/fits-components/Table/FitsTableGroup/type.ts
export interface IFitsTableHeaders extends VxeColumnProps {
    [key: string]: any
}
// ...
export interface IFitsStorageColsSort {
    field?: string
    type?: string
    sort: number
    [key: string]: any
}

/**
 * @desc 存储列显示到localstorage
 */
export interface IFitsStorageColsVisible {
    field?: string
    type?: string
    visible: boolean
    [key: string]: any
}